modify file for solving problem
