+++ 06-building-a-program-start-to-finish/solving-the-problem/Output/inverted-index.js
@@ -1,90 +1,176 @@
-/* eslint-disable no-loop-func*/
-const app = angular.module('inverted_index', []);
-app.controller('myController', ['$scope', ($scope) => {
-  const invertedIndex = new InvertedIndex();
-  $scope.documents = [];
-  $scope.fileNames = [];
-  $scope.numberOfDocuments = {};
-  $scope.indices = {};
-  $scope.showUploaded = false;
-  $scope.showTable = false;
-  $scope.showSearch = false;
-  $scope.file = {};
-  $scope.indexAtr = {};
-  $scope.filesToSearch = [];
-  $scope.fileContent = {};
-  $scope.search = {};
-  const fileUpload = document.getElementById('files');
-  fileUpload.addEventListener('change', () => {
-    $scope.uploadFile(fileUpload);
-  });
-  $scope.uploadFile = (file) => {
-    file = file.files;
-    for (let i = 0; i < file.length; i += 1) {
-      const fileExtension = file[i].name.split('.').pop();
-      if (fileExtension !== 'json') {
-        return toastr.error('This is not a json file');
-      }
-      if ($scope.fileNames.includes(file[i].name)) {
-        return toastr.error(`${file[i].name} has been uploaded`, 'Error');
-      }
-      $scope.documents = $scope.fileNames.push(file[i].name);
-      $scope.$apply();
-      InvertedIndex.readFile(file[i], (e) => {
-        const data = JSON.parse(e.target.result);
-        const fileCheck = InvertedIndex.validateFile(data);
-        if (fileCheck.status) {
-          $scope.fileContent[file[i].name] = data;
-          $scope.$apply();
-          toastr.success(`${file[i].name} uploaded successfully`, 'Success');
-        } else {
-          return toastr.error(fileCheck.msg);
+/**
+ * Implementation of the inverted index data structure.
+ * @author Omokaro Faith <faith.omokaro@andela.com>
+*/
+class InvertedIndex {
+  /**
+   * Creates an instance of InvertedIndex.
+   * @constructor
+   */
+  constructor() {
+    this.index = {};
+    this.temp_search = [];
+    this.searchResult = {};
+    this.allIndex = {};
+    this.token = '';
+  }
+/**
+     * It returns the words passed in small letters and spaces removed
+     * @method tokenize
+     * @param {string} words
+     * @return {String} this.token
+     */
+  tokenize(words) {
+    this.token = words.replace(/,+/g, ' ')
+    .replace(/[^a-zA-Z 0-9\s]+/g, '')
+    .replace(/\s+/g, ' ')
+    .toLowerCase()
+    .trim();
+    return this.token;
+  }
+  /**
+   * Reads the data from an uploaded file.
+   * @param {File} file - The file to be read
+   * @param {Function} callback - The callback function on file read
+   * @returns {void}
+   */
+  static readFile(file, callback) {
+    const reader = new FileReader();
+    reader.onload = callback;
+    reader.readAsText(file);
+  }
+/**
+   * It validates the file passed and returns the error message
+   *
+   * @method validateFile
+   * @param {String} file
+   *
+   * @return {object} validates the file passed
+   */
+  static validateFile(file) {
+    const jsonFile = file;
+    let check = {
+      status: true,
+      msg: 'This is a valid File',
+    };
+    // try {
+    if (!Array.isArray(file) || file.length < 1) {
+      check = {
+        status: false,
+        msg: 'File is empty upload a file',
+      };
+    } else {
+      jsonFile.forEach((key) => {
+        if (key.title === undefined || key.text === undefined) {
+          check = {
+            status: false,
+            msg: 'Invalid file content',
+          };
         }
       });
     }
-  };
-  $scope.createIndex = () => {
-    const fileName = document.getElementById('selectFile').value;
-    const fileContent = $scope.fileContent[fileName];
-    if (fileName.length === 0) {
-      return toastr.error('Upload a file before you create index', 'Error');
-    }
-    try {
-      // fileContent = JSON.parse(fileContent);
-      invertedIndex.createIndex(fileName, fileContent);
-    } catch (err) {
-      toastr.error(err.message);
-    }
-    const index = invertedIndex.getIndex(fileName);
-    $scope.indices[fileName] = index;
-    $scope.fileContent[fileName] = fileContent;
-    const length = fileContent.length;
-    const temp = [];
-    for (let i = 0; i < length; i += 1) {
-      temp.push(i);
-    }
-    $scope.numberOfDocuments[fileName] = temp;
-    $scope.showTable = true;
-    return $scope.filesToSearch.push(fileName);
-  };
+    return check;
+    // } catch (error) {
+    //   return error.msg;
+    // }
+  }
+/**
+     * It returns the index of the file passed
+     *
+     * @method createIndex
+     * @param {Object} fileName
+     * @param {Object} jsonObject
+     *
+     * @return {number} returns the position on the file
+     */
+  createIndex(fileName, jsonObject) {
+    const newIndex = {};
+    // this.validateFile(jsonObject);
 
-  $scope.searchIndex = () => {
-    const filename = document.getElementById('selectSearchFile').value;
-    const words = $scope.searchString;
-    if (filename !== 'Select file') {
-      $scope.searchFileName = filename;
-      $scope.searchResult = invertedIndex.searchIndex(words, filename);
-      $('#modalSearch').modal('show');
-      $('#modalSearchFile').modal('hide');
-    } else {
-      $scope.searchResultAllFiles = invertedIndex.searchIndex(words, filename);
-      $('#modalSearch').modal('hide');
-      $('#modalSearchFile').modal('show');
+    jsonObject.forEach((object, position) => {
+      const longSentence = `${object.title} ${object.text}`;
+      const tokenized = this.tokenize(longSentence);
+      const wordArray = tokenized.split(' ');
+
+      wordArray.forEach((word) => {
+        if (newIndex[word] === undefined) {
+          newIndex[word] = [position];
+        } else if (newIndex[word].indexOf(position) < 0) {
+          newIndex[word].push(position);
+        }
+      });
+    });
+    this.index[fileName] = newIndex;
+  }
+/**
+     * It returns the index of the words
+     * @method getIndex
+     * @param {Object} fileName
+     *
+     * @return {object} gets the index
+     */
+  getIndex(fileName) {
+    if (Object.keys(fileName).length < 1) {
+      return this.index;
     }
-  };
-  $scope.clearSearch = () => {
-    $scope.searchString = '';
-    $('#modalSearch').modal('hide');
-    $('#modalSearchFile').modal('hide');
-  };
-}]);
+
+    return this.index[fileName];
+  }
+  /**
+   * @method flattenSearch
+   * @param {void} arguments of terms
+   * @return {Array} array of terms
+   */
+  flattenSearch(...args) {
+    return args.reduce((acc, val) =>
+      acc.concat(Array.isArray(val) ?
+      this.flattenSearch(val) : val.split(' ')), []);
+  }
+/**
+     * it returns the word searched for in the object it was found
+     *
+     * @method searchIndex
+     * @param {String} sTerms
+     * @param {String} fileName
+     *
+     * @returns {Object} search result
+     */
+  searchIndex(sTerms, fileName) {
+    const searchResult = {};
+    const allIndex = {};
+    if (fileName !== 'Select file') {
+      const selectedIndex = this.index[fileName];
+      let terms = this.tokenize(sTerms);
+      terms = this.flattenSearch(terms);
+      terms.forEach((term) => {
+        if (selectedIndex) {
+          Object.keys(selectedIndex).forEach((savedWord) => {
+            if (savedWord === term) {
+              searchResult[savedWord] = selectedIndex[savedWord];
+            }
+          });
+        }
+      });
+      return searchResult;
+    }
+    Object.keys(this.index).forEach((filename) => {
+      const selectedIndex = this.index[filename];
+      let terms = this.tokenize(sTerms);
+      // this.flattenSearch(terms);
+      terms = this.flattenSearch(terms);
+      terms.forEach((term) => {
+        if (selectedIndex) {
+          Object.keys(selectedIndex).forEach((savedWord) => {
+            if (savedWord === term) {
+              searchResult[savedWord] = selectedIndex[savedWord];
+            }
+          });
+        }
+      });
+      allIndex[filename] = searchResult;
+    });
+    return allIndex;
+  }
+}
+
+module.exports = InvertedIndex;
